verbs: Move CQ related functions into their own file
This moves the CQ related implementation into a
separate file, as a pre-cursor for integrated more
enhanced support for verbs.

As part of this update, the base verbs fid data structures
are moved into a common header.
