+++ Makefile.am
@@ -110,7 +110,10 @@ endif !HAVE_SOCKETS_DL
 endif HAVE_SOCKETS
 
 if HAVE_VERBS
-_verbs_files = prov/verbs/src/fi_verbs.c
+_verbs_files = \
+	prov/verbs/src/fi_verbs.h \
+	prov/verbs/src/fi_verbs.c \
+	prov/verbs/src/verbs_cq.c
 
 if HAVE_VERBS_DL
 pkglib_LTLIBRARIES += libverbs-fi.la
