+++ prov/verbs/src/fi_verbs.c
@@ -34,39 +34,7 @@
 #  include <config.h>
 #endif /* HAVE_CONFIG_H */
 
-#include <asm/types.h>
-#include <errno.h>
-#include <fcntl.h>
-#include <arpa/inet.h>
-#include <netinet/in.h>
-#include <poll.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <unistd.h>
-#include <assert.h>
-#include <pthread.h>
-#include <sys/epoll.h>
-
-#include <infiniband/ib.h>
-#include <infiniband/verbs.h>
-#include <infiniband/driver.h>
-#include <rdma/rdma_cma.h>
-
-#include <rdma/fabric.h>
-#include <rdma/fi_cm.h>
-#include <rdma/fi_domain.h>
-#include <rdma/fi_prov.h>
-#include <rdma/fi_endpoint.h>
-#include <rdma/fi_rma.h>
-#include <rdma/fi_errno.h>
-
-#include "fi.h"
-#include "fi_enosys.h"
-#include <rdma/fi_log.h>
-#include "prov.h"
-#include "fi_list.h"
-#include "fi_signal.h"
+#include "fi_verbs.h"
 
 
 static int fi_ibv_getinfo(uint32_t version, const char *node, const char *service,
@@ -78,7 +46,7 @@ static void fi_ibv_fini(void);
 #define VERBS_PROV_NAME "verbs"
 #define VERBS_PROV_VERS FI_VERSION(1,0)
 
-static struct fi_provider fi_ibv_prov = {
+struct fi_provider fi_ibv_prov = {
 	.name = VERBS_PROV_NAME,
 	.version = VERBS_PROV_VERS,
 	.fi_version = FI_VERSION(1, 1),
@@ -121,11 +89,6 @@ static struct fi_provider fi_ibv_prov = {
 
 #define VERBS_COMP_READ(ep) VERBS_COMP_READ_FLAGS(ep, ep->info->tx_attr->op_flags)
 
-#define VERBS_DBG(subsys, ...) FI_DBG(&fi_ibv_prov, subsys, __VA_ARGS__)
-#define VERBS_INFO(subsys, ...) FI_INFO(&fi_ibv_prov, subsys, __VA_ARGS__)
-
-#define VERBS_INFO_ERRNO(subsys, fn, errno) VERBS_INFO(subsys, fn ": %s(%d)\n",	\
-		strerror(errno), errno)
 
 #define fi_ibv_set_sge(sge, buf, len, desc)		\
 	do {						\
@@ -169,75 +132,6 @@ static struct fi_provider fi_ibv_prov = {
 		}							\
 	} while (0)
 
-struct fi_ibv_fabric {
-	struct fid_fabric	fabric_fid;
-};
-
-struct fi_ibv_eq_entry {
-	struct dlist_entry	item;
-	uint32_t		event;
-	size_t			len;
-	char 			eq_entry[0];
-};
-
-struct fi_ibv_eq {
-	struct fid_eq		eq_fid;
-	struct fi_ibv_fabric	*fab;
-	fastlock_t		lock;
-	struct dlistfd_head	list_head;
-	struct rdma_event_channel *channel;
-	uint64_t		flags;
-	struct fi_eq_err_entry	err;
-	int			epfd;
-};
-
-struct fi_ibv_pep {
-	struct fid_pep		pep_fid;
-	struct fi_ibv_eq	*eq;
-	struct rdma_cm_id	*id;
-	int			backlog;
-	int			bound;
-	size_t			src_addrlen;
-};
-
-struct fi_ibv_domain {
-	struct fid_domain	domain_fid;
-	struct ibv_context	*verbs;
-	struct ibv_pd		*pd;
-};
-
-struct fi_ibv_cq {
-	struct fid_cq		cq_fid;
-	struct fi_ibv_domain	*domain;
-	struct ibv_comp_channel	*channel;
-	struct ibv_cq		*cq;
-	size_t			entry_size;
-	uint64_t		flags;
-	enum fi_cq_wait_cond	wait_cond;
-	struct ibv_wc		wc;
-	int			signal_fd[2];
-};
-
-struct fi_ibv_mem_desc {
-	struct fid_mr		mr_fid;
-	struct ibv_mr		*mr;
-	struct fi_ibv_domain	*domain;
-};
-
-struct fi_ibv_msg_ep {
-	struct fid_ep		ep_fid;
-	struct rdma_cm_id	*id;
-	struct fi_ibv_eq	*eq;
-	struct fi_ibv_cq	*rcq;
-	struct fi_ibv_cq	*scq;
-	uint64_t		ep_flags;
-	struct fi_info		*info;
-};
-
-struct fi_ibv_connreq {
-	struct fid		handle;
-	struct rdma_cm_id	*id;
-};
 
 static const char *local_node = "localhost";
 static char def_tx_ctx_size[16] = "384";
@@ -2917,459 +2811,6 @@ err2:
 err1:
 	fastlock_destroy(&_eq->lock);
 	free(_eq);
-	return ret;
-}
-
-static ssize_t
-fi_ibv_cq_readerr(struct fid_cq *cq, struct fi_cq_err_entry *entry,
-		  uint64_t flags)
-{
-	struct fi_ibv_cq *_cq;
-
-	_cq = container_of(cq, struct fi_ibv_cq, cq_fid);
-	if (!_cq->wc.status)
-		return 0;
-
-	entry->op_context = (void *) (uintptr_t) _cq->wc.wr_id;
-	entry->flags = 0;
-	entry->err = EIO;
-	entry->prov_errno = _cq->wc.status;
-	memcpy(&entry->err_data, &_cq->wc.vendor_err,
-	       sizeof(_cq->wc.vendor_err));
-
-	_cq->wc.status = 0;
-	return sizeof(*entry);
-}
-
-static inline int
-fi_ibv_poll_events(struct fi_ibv_cq *_cq, int timeout)
-{
-	int ret;
-	void *context;
-	struct pollfd fds[2];
-	char data;
-
-	fds[0].fd = _cq->channel->fd;
-	fds[1].fd = _cq->signal_fd[0];
-
-	fds[0].events = fds[1].events = POLLIN;
-
-	ret = poll(fds, 2, timeout);
-	if (ret == 0)
-		return -FI_EAGAIN;
-	else if (ret < 0)
-		return -errno;
-
-	if (fds[1].revents & POLLIN) {
-		do {
-			ret = read(fds[1].fd, &data, 1);
-		} while (ret > 0);
-		return -FI_EAGAIN;
-	} else if (fds[0].revents & POLLIN) {
-		ret = ibv_get_cq_event(_cq->channel, &_cq->cq, &context);
-		if (ret)
-			return ret;
-
-		ibv_ack_cq_events(_cq->cq, 1);
-	} else {
-		FI_WARN(&fi_ibv_prov, FI_LOG_CQ, "Unknown poll error: check revents\n");
-		return -FI_EOTHER;
-	}
-
-	return 0;
-}
-
-static ssize_t
-fi_ibv_cq_sread(struct fid_cq *cq, void *buf, size_t count, const void *cond,
-		int timeout)
-{
-	ssize_t ret = 0, cur;
-	ssize_t  threshold;
-	struct fi_ibv_cq *_cq;
-
-	_cq = container_of(cq, struct fi_ibv_cq, cq_fid);
-
-	if (!_cq->channel)
-		return -FI_ENOSYS;
-
-	threshold = (_cq->wait_cond == FI_CQ_COND_THRESHOLD) ?
-		MIN((ssize_t) cond, count) : 1;
-
-	for (cur = 0; cur < threshold; ) {
-		ret = _cq->cq_fid.ops->read(&_cq->cq_fid, buf, count - cur);
-		if (ret > 0) {
-			buf += ret * _cq->entry_size;
-			cur += ret;
-			if (cur >= threshold)
-				break;
-		} else if (ret != -FI_EAGAIN) {
-			break;
-		}
-
-		ret = ibv_req_notify_cq(_cq->cq, 0);
-		if (ret) {
-			FI_WARN(&fi_ibv_prov, FI_LOG_CQ, "ibv_req_notify_cq error: %d\n", ret);
-			break;
-		}
-
-		/* Read again to fetch any completions that we might have missed
-		 * while rearming */
-		ret = _cq->cq_fid.ops->read(&_cq->cq_fid, buf, count - cur);
-		if (ret > 0) {
-			buf += ret * _cq->entry_size;
-			cur += ret;
-			if (cur >= threshold)
-				break;
-		} else if (ret != -FI_EAGAIN) {
-			break;
-		}
-
-		ret = fi_ibv_poll_events(_cq, timeout);
-		if (ret)
-			break;
-	}
-
-	return cur ? cur : ret;
-}
-
-static uint64_t fi_ibv_comp_flags(struct ibv_wc *wc)
-{
-	uint64_t flags = 0;
-
-	if (wc->wc_flags & IBV_WC_WITH_IMM)
-		flags |= FI_REMOTE_CQ_DATA;
-
-	switch (wc->opcode) {
-	case IBV_WC_SEND:
-		flags |= FI_SEND | FI_MSG;
-		break;
-	case IBV_WC_RDMA_WRITE:
-		flags |= FI_RMA | FI_WRITE;
-		break;
-	case IBV_WC_RDMA_READ:
-		flags |= FI_RMA | FI_READ;
-		break;
-	case IBV_WC_COMP_SWAP:
-		flags |= FI_ATOMIC;
-		break;
-	case IBV_WC_FETCH_ADD:
-		flags |= FI_ATOMIC;
-		break;
-	case IBV_WC_RECV:
-		flags |= FI_RECV | FI_MSG;
-		break;
-	case IBV_WC_RECV_RDMA_WITH_IMM:
-		flags |= FI_RMA | FI_REMOTE_WRITE;
-		break;
-	default:
-		break;
-	}
-	return flags;
-}
-
-static ssize_t fi_ibv_cq_read_context(struct fid_cq *cq, void *buf, size_t count)
-{
-	struct fi_ibv_cq *_cq;
-	struct fi_cq_entry *entry = buf;
-	ssize_t ret = 0, i;
-
-	_cq = container_of(cq, struct fi_ibv_cq, cq_fid);
-	if (_cq->wc.status)
-		return -FI_EAVAIL;
-
-	for (i = 0; i < count; i++) {
-		ret = ibv_poll_cq(_cq->cq, 1, &_cq->wc);
-		if (ret <= 0)
-			break;
-
-		if (_cq->wc.status) {
-			ret = -FI_EAVAIL;
-			break;
-		}
-
-		entry->op_context = (void *) (uintptr_t) _cq->wc.wr_id;
-		entry += 1;
-	}
-
-	return i ? i : (ret ? ret : -FI_EAGAIN);
-}
-
-static ssize_t fi_ibv_cq_read_msg(struct fid_cq *cq, void *buf, size_t count)
-{
-	struct fi_ibv_cq *_cq;
-	struct fi_cq_msg_entry *entry = buf;
-	ssize_t ret = 0, i;
-
-	_cq = container_of(cq, struct fi_ibv_cq, cq_fid);
-	if (_cq->wc.status)
-		return -FI_EAVAIL;
-
-	for (i = 0; i < count; i++) {
-		ret = ibv_poll_cq(_cq->cq, 1, &_cq->wc);
-		if (ret <= 0)
-			break;
-
-		if (_cq->wc.status) {
-			ret = -FI_EAVAIL;
-			break;
-		}
-
-		entry->op_context = (void *) (uintptr_t) _cq->wc.wr_id;
-		entry->flags = fi_ibv_comp_flags(&_cq->wc);
-		entry->len = (uint64_t) _cq->wc.byte_len;
-		entry += 1;
-	}
-
-	return i ? i : (ret ? ret : -FI_EAGAIN);
-}
-
-static ssize_t fi_ibv_cq_read_data(struct fid_cq *cq, void *buf, size_t count)
-{
-	struct fi_ibv_cq *_cq;
-	struct fi_cq_data_entry *entry = buf;
-	ssize_t ret = 0, i;
-
-	_cq = container_of(cq, struct fi_ibv_cq, cq_fid);
-	if (_cq->wc.status)
-		return -FI_EAVAIL;
-
-	for (i = 0; i < count; i++) {
-		ret = ibv_poll_cq(_cq->cq, 1, &_cq->wc);
-		if (ret <= 0)
-			break;
-
-		if (_cq->wc.status) {
-			ret = -FI_EAVAIL;
-			break;
-		}
-
-		entry->op_context = (void *) (uintptr_t) _cq->wc.wr_id;
-		entry->flags = fi_ibv_comp_flags(&_cq->wc);
-		if (_cq->wc.wc_flags & IBV_WC_WITH_IMM) {
-			entry->data = ntohl(_cq->wc.imm_data);
-		} else {
-			entry->data = 0;
-		}
-		if (_cq->wc.opcode & (IBV_WC_RECV | IBV_WC_RECV_RDMA_WITH_IMM))
-			entry->len = _cq->wc.byte_len;
-		else
-			entry->len = 0;
-
-		entry += 1;
-	}
-
-	return i ? i : (ret ? ret : -FI_EAGAIN);
-}
-
-static const char *
-fi_ibv_cq_strerror(struct fid_cq *eq, int prov_errno, const void *err_data,
-		   char *buf, size_t len)
-{
-	if (buf && len)
-		strncpy(buf, ibv_wc_status_str(prov_errno), len);
-	return ibv_wc_status_str(prov_errno);
-}
-
-static int
-fi_ibv_cq_signal(struct fid_cq *cq)
-{
-	struct fi_ibv_cq *_cq;
-	char data = '0';
-
-	_cq = container_of(cq, struct fi_ibv_cq, cq_fid);
-
-	if (write(_cq->signal_fd[1], &data, 1) != 1) {
-		FI_WARN(&fi_ibv_prov, FI_LOG_CQ, "Error signalling CQ\n");
-		return -errno;
-	}
-
-	return 0;
-}
-
-static struct fi_ops_cq fi_ibv_cq_context_ops = {
-	.size = sizeof(struct fi_ops_cq),
-	.read = fi_ibv_cq_read_context,
-	.readfrom = fi_no_cq_readfrom,
-	.readerr = fi_ibv_cq_readerr,
-	.sread = fi_ibv_cq_sread,
-	.sreadfrom = fi_no_cq_sreadfrom,
-	.signal = fi_ibv_cq_signal,
-	.strerror = fi_ibv_cq_strerror
-};
-
-static struct fi_ops_cq fi_ibv_cq_msg_ops = {
-	.size = sizeof(struct fi_ops_cq),
-	.read = fi_ibv_cq_read_msg,
-	.readfrom = fi_no_cq_readfrom,
-	.readerr = fi_ibv_cq_readerr,
-	.sread = fi_ibv_cq_sread,
-	.sreadfrom = fi_no_cq_sreadfrom,
-	.signal = fi_ibv_cq_signal,
-	.strerror = fi_ibv_cq_strerror
-};
-
-static struct fi_ops_cq fi_ibv_cq_data_ops = {
-	.size = sizeof(struct fi_ops_cq),
-	.read = fi_ibv_cq_read_data,
-	.readfrom = fi_no_cq_readfrom,
-	.readerr = fi_ibv_cq_readerr,
-	.sread = fi_ibv_cq_sread,
-	.sreadfrom = fi_no_cq_sreadfrom,
-	.signal = fi_ibv_cq_signal,
-	.strerror = fi_ibv_cq_strerror
-};
-
-static int fi_ibv_cq_control(fid_t fid, int command, void *arg)
-{
-	struct fi_ibv_cq *cq;
-	int ret = 0;
-
-	cq = container_of(fid, struct fi_ibv_cq, cq_fid.fid);
-	switch(command) {
-	case FI_GETWAIT:
-		if (!cq->channel) {
-			ret = -FI_ENODATA;
-			break;
-		}
-		*(int *) arg = cq->channel->fd;
-		break;
-	default:
-		ret = -FI_ENOSYS;
-		break;
-	}
-
-	return ret;
-}
-
-static int fi_ibv_cq_close(fid_t fid)
-{
-	struct fi_ibv_cq *cq;
-	int ret;
-
-	cq = container_of(fid, struct fi_ibv_cq, cq_fid.fid);
-	if (cq->cq) {
-		ret = ibv_destroy_cq(cq->cq);
-		if (ret)
-			return -ret;
-	}
-
-	if (cq->signal_fd[0]) {
-		close(cq->signal_fd[0]);
-	}
-	if (cq->signal_fd[1]) {
-		close(cq->signal_fd[1]);
-	}
-
-	if (cq->channel)
-		ibv_destroy_comp_channel(cq->channel);
-
-	free(cq);
-	return 0;
-}
-
-static struct fi_ops fi_ibv_cq_fi_ops = {
-	.size = sizeof(struct fi_ops),
-	.close = fi_ibv_cq_close,
-	.bind = fi_no_bind,
-	.control = fi_ibv_cq_control,
-	.ops_open = fi_no_ops_open,
-};
-
-static int
-fi_ibv_cq_open(struct fid_domain *domain, struct fi_cq_attr *attr,
-	    struct fid_cq **cq, void *context)
-{
-	struct fi_ibv_cq *_cq;
-	int ret;
-
-	_cq = calloc(1, sizeof *_cq);
-	if (!_cq)
-		return -FI_ENOMEM;
-
-	_cq->domain = container_of(domain, struct fi_ibv_domain, domain_fid);
-
-	switch (attr->wait_obj) {
-	case FI_WAIT_UNSPEC:
-	case FI_WAIT_FD:
-		_cq->channel = ibv_create_comp_channel(_cq->domain->verbs);
-		if (!_cq->channel) {
-			ret = -errno;
-			goto err1;
-		}
-
-		ret = fi_fd_nonblock(_cq->channel->fd);
-		if (ret)
-			goto err2;
-
-		if (socketpair(AF_UNIX, SOCK_STREAM, 0, _cq->signal_fd)) {
-			ret = -errno;
-			goto err2;
-		}
-
-		ret = fi_fd_nonblock(_cq->signal_fd[0]);
-		if (ret)
-			goto err3;
-
-		break;
-	case FI_WAIT_NONE:
-		break;
-	default:
-		ret = -FI_ENOSYS;
-		goto err3;
-	}
-
-	_cq->cq = ibv_create_cq(_cq->domain->verbs, attr->size, _cq,
-				_cq->channel, attr->signaling_vector);
-	if (!_cq->cq) {
-		ret = -errno;
-		goto err3;
-	}
-
-	if (_cq->channel) {
-		ret = ibv_req_notify_cq(_cq->cq, 0);
-		if (ret) {
-			FI_WARN(&fi_ibv_prov, FI_LOG_CQ, "ibv_req_notify_cq failed\n");
-			goto err4;
-		}
-	}
-
-	_cq->flags |= attr->flags;
-	_cq->wait_cond = attr->wait_cond;
-	_cq->cq_fid.fid.fclass = FI_CLASS_CQ;
-	_cq->cq_fid.fid.context = context;
-	_cq->cq_fid.fid.ops = &fi_ibv_cq_fi_ops;
-
-	switch (attr->format) {
-	case FI_CQ_FORMAT_CONTEXT:
-		_cq->cq_fid.ops = &fi_ibv_cq_context_ops;
-		_cq->entry_size = sizeof(struct fi_cq_entry);
-		break;
-	case FI_CQ_FORMAT_MSG:
-		_cq->cq_fid.ops = &fi_ibv_cq_msg_ops;
-		_cq->entry_size = sizeof(struct fi_cq_msg_entry);
-		break;
-	case FI_CQ_FORMAT_DATA:
-		_cq->cq_fid.ops = &fi_ibv_cq_data_ops;
-		_cq->entry_size = sizeof(struct fi_cq_data_entry);
-		break;
-	default:
-		ret = -FI_ENOSYS;
-		goto err4;
-	}
-
-	*cq = &_cq->cq_fid;
-	return 0;
-
-err4:
-	ibv_destroy_cq(_cq->cq);
-err3:
-	close(_cq->signal_fd[0]);
-	close(_cq->signal_fd[1]);
-err2:
-	if (_cq->channel)
-		ibv_destroy_comp_channel(_cq->channel);
-err1:
-	free(_cq);
 	return ret;
 }
