+++ prov/verbs/src/fi_verbs.h
@@ -1,0 +1,154 @@
+/*
+ * Copyright (c) 2013-2015 Intel Corporation, Inc.  All rights reserved.
+ *
+ * This software is available to you under a choice of one of two
+ * licenses.  You may choose to be licensed under the terms of the GNU
+ * General Public License (GPL) Version 2, available from the file
+ * COPYING in the main directory of this source tree, or the
+ * BSD license below:
+ *
+ *     Redistribution and use in source and binary forms, with or
+ *     without modification, are permitted provided that the following
+ *     conditions are met:
+ *
+ *      - Redistributions of source code must retain the above
+ *        copyright notice, this list of conditions and the following
+ *        disclaimer.
+ *
+ *      - Redistributions in binary form must reproduce the above
+ *        copyright notice, this list of conditions and the following
+ *        disclaimer in the documentation and/or other materials
+ *        provided with the distribution.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS
+ * BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
+ * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+#if HAVE_CONFIG_H
+#  include <config.h>
+#endif /* HAVE_CONFIG_H */
+
+#include <asm/types.h>
+#include <errno.h>
+#include <fcntl.h>
+#include <arpa/inet.h>
+#include <netinet/in.h>
+#include <poll.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+#include <assert.h>
+#include <pthread.h>
+#include <sys/epoll.h>
+
+#include <infiniband/ib.h>
+#include <infiniband/verbs.h>
+#include <infiniband/driver.h>
+#include <rdma/rdma_cma.h>
+
+#include <rdma/fabric.h>
+#include <rdma/fi_cm.h>
+#include <rdma/fi_domain.h>
+#include <rdma/fi_prov.h>
+#include <rdma/fi_endpoint.h>
+#include <rdma/fi_rma.h>
+#include <rdma/fi_errno.h>
+
+#include "fi.h"
+#include "fi_enosys.h"
+#include <rdma/fi_log.h>
+#include "prov.h"
+#include "fi_list.h"
+#include "fi_signal.h"
+
+
+#define VERBS_DBG(subsys, ...) FI_DBG(&fi_ibv_prov, subsys, __VA_ARGS__)
+#define VERBS_INFO(subsys, ...) FI_INFO(&fi_ibv_prov, subsys, __VA_ARGS__)
+
+#define VERBS_INFO_ERRNO(subsys, fn, errno) VERBS_INFO(subsys, fn ": %s(%d)\n",	\
+		strerror(errno), errno)
+
+
+extern struct fi_provider fi_ibv_prov;
+
+
+struct fi_ibv_fabric {
+	struct fid_fabric	fabric_fid;
+};
+
+struct fi_ibv_eq_entry {
+	struct dlist_entry	item;
+	uint32_t		event;
+	size_t			len;
+	char 			eq_entry[0];
+};
+
+struct fi_ibv_eq {
+	struct fid_eq		eq_fid;
+	struct fi_ibv_fabric	*fab;
+	fastlock_t		lock;
+	struct dlistfd_head	list_head;
+	struct rdma_event_channel *channel;
+	uint64_t		flags;
+	struct fi_eq_err_entry	err;
+	int			epfd;
+};
+
+struct fi_ibv_pep {
+	struct fid_pep		pep_fid;
+	struct fi_ibv_eq	*eq;
+	struct rdma_cm_id	*id;
+	int			backlog;
+	int			bound;
+	size_t			src_addrlen;
+};
+
+struct fi_ibv_domain {
+	struct fid_domain	domain_fid;
+	struct ibv_context	*verbs;
+	struct ibv_pd		*pd;
+};
+
+struct fi_ibv_cq {
+	struct fid_cq		cq_fid;
+	struct fi_ibv_domain	*domain;
+	struct ibv_comp_channel	*channel;
+	struct ibv_cq		*cq;
+	size_t			entry_size;
+	uint64_t		flags;
+	enum fi_cq_wait_cond	wait_cond;
+	struct ibv_wc		wc;
+	int			signal_fd[2];
+};
+
+int fi_ibv_cq_open(struct fid_domain *domain, struct fi_cq_attr *attr,
+		   struct fid_cq **cq, void *context);
+
+
+struct fi_ibv_mem_desc {
+	struct fid_mr		mr_fid;
+	struct ibv_mr		*mr;
+	struct fi_ibv_domain	*domain;
+};
+
+struct fi_ibv_msg_ep {
+	struct fid_ep		ep_fid;
+	struct rdma_cm_id	*id;
+	struct fi_ibv_eq	*eq;
+	struct fi_ibv_cq	*rcq;
+	struct fi_ibv_cq	*scq;
+	uint64_t		ep_flags;
+	struct fi_info		*info;
+};
+
+struct fi_ibv_connreq {
+	struct fid		handle;
+	struct rdma_cm_id	*id;
+};
